% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetictree.R
\name{phylogenetictree}
\alias{phylogenetictree}
\title{<Add Title>}
\usage{
phylogenetictree(tree, outerradius = 480, innerradius = 310, width = NULL,
  height = NULL)
}
\description{
<Add Description>
}

