% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetictree.R
\name{phylogenetictreeOutput}
\alias{phylogenetictreeOutput}
\title{Widget output function for use in Shiny}
\usage{
phylogenetictreeOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

