% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetictree.R
\name{renderPhylogenetictree}
\alias{renderPhylogenetictree}
\title{Widget render function for use in Shiny}
\usage{
renderPhylogenetictree(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

